% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cluspred.R
\docType{package}
\name{ClusPred-package}
\alias{ClusPred-package}
\alias{ClusPred}
\title{ClusPred.}
\description{
Parameter estimation of regression models with fixed group effects, when the group variable is missing while group-related variables are available.
}
\details{
\tabular{ll}{
  Package: \tab ClusPred\cr
  Type: \tab Package\cr
  Version: \tab 1.0.0\cr
  Date: \tab 2021-06-01\cr
  License: \tab GPL-3\cr
  LazyLoad: \tab yes\cr
}
}
\references{
Simultaneous semi-parametric estimation of clustering and regression, Matthieu Marbac and Mohammed Sedki and Christophe Biernacki and Vincent Vandewalle (2020) <arXiv:2012.14159>.
}
\keyword{package}
